Refactored code to reuse existing code
